requirements.txt
neat-python==0.92
numpy==1.22.2
pandas==1.4.1
pygame==2.0.2
Pillow==9.2.0
sympy==1.11.1